feat: BaseViewModelFactory로 추상화하여 팩토리 메서드에서 사용
