+++ app/src/main/java/woowacourse/shopping/presentation/ui/productdetail/ProductDetailActivity.kt
@@ -14,7 +14,8 @@ import woowacourse.shopping.presentation.base.observeEvent
 
 class ProductDetailActivity : BaseActivity<ActivityProductDetailBinding>() {
     private val viewModel: ProductDetailViewModel by viewModels {
-        ProductDetailViewModelFactory(
+        ProductDetailViewModel.factory(
+            defaultViewModelCreationExtras,
             (application as App).productRepository,
             (application as App).shoppingCartRepository,
         )
