+++ app/src/main/java/woowacourse/shopping/presentation/ui/productdetail/ProductDetailViewModel.kt
@@ -3,10 +3,14 @@ package woowacourse.shopping.presentation.ui.productdetail
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.createSavedStateHandle
+import androidx.lifecycle.viewmodel.CreationExtras
 import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.ShoppingCartRepository
 import woowacourse.shopping.presentation.base.BaseViewModel
+import woowacourse.shopping.presentation.base.BaseViewModelFactory
 import woowacourse.shopping.presentation.base.MessageProvider
 import woowacourse.shopping.presentation.ui.productdetail.ProductDetailActivity.Companion.PUT_EXTRA_PRODUCT_ID
 
@@ -39,6 +43,22 @@ class ProductDetailViewModel(
         product.value?.let { product ->
             shoppingCartRepository.addOrder(product)
             showMessage(ProductDetailMessage.AddToCartSuccessMessage)
+        }
+    }
+
+    companion object {
+        fun factory(
+            extras: CreationExtras,
+            productRepository: ProductRepository,
+            shoppingCartRepository: ShoppingCartRepository,
+        ): ViewModelProvider.Factory {
+            return BaseViewModelFactory {
+                ProductDetailViewModel(
+                    extras.createSavedStateHandle(),
+                    productRepository,
+                    shoppingCartRepository,
+                )
+            }
         }
     }
 }
